# WebApi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MyBlog API",
    "version": "1.0.0",
    "description": "Auto-generated OpenAPI JSON for MyBlog WebApi project"
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Local server" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ArticleDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "content": { "type": "string" },
          "author": { "type": "string" },
          "categoryId": { "type": "string", "format": "uuid" },
          "createdDate": { "type": "string" },
          "updatedDate": { "type": ["string", "null"] },
          "isDeleted": { "type": "boolean" },
          "deletedDate": { "type": ["string", "null"] }
        }
      },
      "CreateArticleDto": {
        "type": "object",
        "required": ["Content", "Title", "Author", "CategoryId"],
        "properties": {
          "Content": { "type": "string" },
          "Title": { "type": "string" },
          "Author": { "type": "string" },
          "CategoryId": { "type": "string", "format": "uuid" }
        }
      },
      "UpdateArticleDto": { "$ref": "#/components/schemas/CreateArticleDto" },
...
      "ResponseModel_Generic": {
        "type": "object",
        "properties": {
          "isSuccess": { "type": "boolean" },
          "failMessages": { "type": "array", "items": { "type": "string" } },
          "data": { }
        }
      },
      "ResponseModel_ArticleArray": {
        "allOf": [
          { "$ref": "#/components/schemas/ResponseModel_Generic" },
          { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/ArticleDto" } } } }
        ]
      },
      "ResponseModel_Article": {
        "allOf": [
          { "$ref": "#/components/schemas/ResponseModel_Generic" },
          { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/ArticleDto" } } }
        ]
      },
      "ResponseModel_Guid": {
        "allOf": [
          { "$ref": "#/components/schemas/ResponseModel_Generic" },
          { "type": "object", "properties": { "data": { "type": "string", "format": "uuid" } } }
        ]
      },
...
    }
  },
  "paths": {
    "/api/articles": {
      "get": {
        "summary": "Get all articles",
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseModel_ArticleArray" } } } } }
      },
      "post": {
        "summary": "Create article",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateArticleDto" } } }
        },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseModel_Guid" } } } } }
      }
    },
    "/api/articles/{id}": {
      "get": {
        "summary": "Get article by id",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseModel_Article" } } } } }
      },
      "put": {
        "summary": "Update article (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateArticleDto" } } } },
        "responses": { "204": { "description": "No Content", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseModel_Generic" } } } } }
      },
      "delete": {
        "summary": "Delete article (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "204": { "description": "No Content", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseModel_Generic" } } } } }
      }
    },
    "/api/articles/getpaginate": {
      "get": {
        "summary": "Get articles by pagination",
        "parameters": [
          { "name": "take", "in": "query", "required": true, "schema": { "type": "integer" } },
          { "name": "skip", "in": "query", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseModel_ArticleArray" } } } } }
      }
    },
...
  }
}